Install runs

+++ hoServices/src/AddInSearch/NestedObject.cs
@@ -4,15 +4,15 @@ namespace EaServices.AddInSearch
 {
     public class NestedObject
     {
-        public int ParentId;
-        public int TPos;
+        public long ParentId;
+        public long TPos;
         public string Name;
         public string Alias;
         public string Notes;
         public List<Tv> Tv;
         public string Guid;
         public string ObjectType;
-        public NestedObject(int parentId, int tPos, string name, string alias, string notes, string guid, string objectType,  List<Tv> tv)
+        public NestedObject(long parentId, long tPos, string name, string alias, string notes, string guid, string objectType,  List<Tv> tv)
         {
             ParentId = parentId;
             TPos = tPos;
@@ -24,5 +24,6 @@ namespace EaServices.AddInSearch
             Tv = tv;
 
         }
+        
     }
 }
